// apps/portal/src/api/generated/model/unitOwnershipPeriodResponseDto.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * HOA Manager API
 * The HOA Manager API description
 * OpenAPI spec version: 1.0
 */
import type { UnitOwnershipPeriodResponseDtoStatus } from "./unitOwnershipPeriodResponseDtoStatus";
import type { UnitOwnershipResponseDto } from "./unitOwnershipResponseDto";

export interface UnitOwnershipPeriodResponseDto {
    validFrom: string;
    /** @nullable */
    validTo: string | null;
    status: UnitOwnershipPeriodResponseDtoStatus;
    parties: UnitOwnershipResponseDto[];
}